+++ ui/src/DataService.js
@@ -65,7 +65,6 @@ const webSocket = (action, url) => {
 const dataService = {
   mobileClients: () => fetchItems('mobileclients'),
   serviceInstances: () => fetchItems('serviceinstances'),
-  bindableServices: () => fetchItems('bindableservices'),
   builds: () => fetchItems('builds'),
   buildConfigs: () => fetchItems('buildconfigs'),
   createApp: app => request('mobileclients', 'POST', app),
@@ -75,7 +74,6 @@ const dataService = {
   deleteBuildConfig: name => deleteItem(`buildconfigs/${name}`, name),
   updateApp: async (id, app) => {
     const response = await fetch(`${baseUrl}/mobileclients/${id}`, {
-
       method: 'POST',
       cache: 'no-cache',
       credentials: 'same-origin',
@@ -89,9 +87,58 @@ const dataService = {
       throw Error(`${response.statusText}: ${msg}`);
     }
     return response.json();
-
   },
 
+  bindableServices: async (mobileClientName) => {
+    fetchItems(`${baseUrl}/bindableservices/${mobileClientName}`)
+    .then(instances => {
+            let unboundServices = [];
+            let boundServices = [];
+            let state = {};
+             instances.forEach ( instance => {
+               
+                let serviceName = instance.name;
+                let serviceIcon = instance.imageUrl;
+                let serviceIconClass = instance.iconClass;
+                
+                if (instance.isBound) {
+                  
+                  boundServices.push({
+                    serviceLogoUrl: serviceIcon,
+                    serviceIconClass: serviceIconClass,
+                    serviceName: serviceName,
+                    serviceBindingName: instance.serviceBinding.metadata.name,
+                    serviceInstanceName: instance.serviceInstance.metadata.name,
+                    serviceId: serviceName,
+                    serviceDescription: instance.serviceClass.spec.description,
+                    documentationUrl: instance.serviceClass.spec.externalMetadata.documentationUrl20,
+                    configuration: instance.configuration,
+                    setupText: 'Identity Management SDK setup',            
+                  });
+                } else {
+                  unboundServices.push({
+                    serviceLogoUrl: serviceIcon,
+                    serviceIconClass: serviceIconClass,
+                    serviceName: serviceName,
+                    serviceInstanceName: instance.serviceInstance.metadata.name,
+                    serviceId: serviceName,
+                    bindingSchema : instance.servicePlan.spec.serviceBindingCreateParameterSchema,
+                    form : instance.servicePlan.spec.externalMetadata.schemas.service_binding.create.openshift_form_definition,
+                    serviceDescription: instance.serviceClass.spec.description,
+                    serviceClassExternalName: instance.serviceClass.spec.externalMetadata.serviceName,
+                    setupText: 'Mobile Metrics SDK setups',
+                  });
+                }
+                
+                state.boundServices = boundServices;
+                state.unboundServices = unboundServices;
+                return state;
+            })
+          }
+        );
+      
+  },
+  
   createBinding: async (mobileClientName, serviceInstanceName, credentialSecretName, parametersSecretName, serviceClassExternalName, formData) => {
 
     var binding = {
