+++ ui/src/components/overview/MobileClientCardView.js
@@ -24,6 +24,7 @@ class MobileClientCardView extends Component {
   constructor(props) {
     super(props);
     this.state = { filter: '', currentValue: '' };
+    
     this.emptyStateMessage = {
       noAppsCreated: 'You have no mobile apps right now. Create one to get started.',
       noAppsAfterFiltering: 'No mobile apps match the entered filter.'
